Updated API endpoints to new structure following #25
