+++ pages/api/check-domain.ts
@@ -4,6 +4,11 @@ export default async function checkDomain(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  if (req.method !== "GET") {
+    res.setHeader("Allow", ["GET"]);
+    return res.status(405).end(`Method ${req.method} Not Allowed`);
+  }
+
   const { domain } = req.query;
 
   try {
